Add files via upload
This is my first small project in php .A php CRUD (Create, Read, Update, Delete) project is a common type of web application that demonstrates the basic operations of managing data in a database. CRUD operations are the fundamental actions for interacting with a database, allowing users to create new records, retrieve existing records, update record information, and delete records as needed.